v2 - remove redundant media query items
Because `media-breakpoint-up` is used to define gutters, we can check to see if the previous gutter is the same of the current one.  If so, don't add a media query.

A few grid tweaks to reduce reduncancy and fix some missing/incorrect rule creation.

Also fix the base `$grid-gutter-width` to be the documented 2rem.