Remove focus `box-shadow` from btn and form controls
It seems, at least personally, that I am almost always turning this off and adding a focus/hover color.
Visually it is great when put on a light color background, such as the default body background color, but when used on a dark background, they just dont seem to work well.

So we are going back to having a consolidated hover/focus coloring.