Offcanvas: handle resize to reset scrollbar and backdrop (#809)
Stops rootElement scroll on mobile when offcanvas is visually blocking the entire viewport.

Hides (resets backdrop and scrollbar) for an open offcanvas when no longer positioned (fixed/absolute), such as when used in an 'expanded'  navbar.  When in a navbar the offcanvas switches from positioned when open to non-positioned and always visible as 'inline' content.